refactor: clean up style imports
